+++ domestic/models.py
@@ -232,6 +232,16 @@ class GreatDomesticHomePage(
 
     # EU exit chevrons StreamField WAS here in V1 - no longer the case
 
+    # magna ctas
+    magna_ctas_title = models.TextField(null=True, blank=True)
+    magna_ctas_columns = single_struct_block_stream_field_factory(
+        field_name='columns',
+        block_class_instance=core_blocks.LinkWithImageAndContentBlockNoSource(),
+        max_num=3,
+        null=True,
+        blank=True,
+    )
+
     # how DIT helps
     how_dit_helps_title = models.TextField(null=True, blank=True)
     how_dit_helps_columns = single_struct_block_stream_field_factory(